added product and user compliance check

+++ product-service/src/main/java/com/alexandersuetnov/productservice/service/ProductService.java
@@ -1,5 +1,6 @@
 package com.alexandersuetnov.productservice.service;
 
+import com.alexandersuetnov.productservice.config.JWTFilter;
 import com.alexandersuetnov.productservice.dto.ProductDTO;
 import com.alexandersuetnov.productservice.exception.ProductNotFoundException;
 import com.alexandersuetnov.productservice.mappers.ProductMapper;
@@ -9,6 +10,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.stereotype.Service;
 
+import javax.servlet.http.HttpServletRequest;
 import java.security.Principal;
 import java.util.List;
 
@@ -18,15 +20,16 @@ import java.util.List;
 public class ProductService {
 
     private final ProductRepository productRepository;
+    private final JWTFilter jwtFilter;
 
     public Product createProduct(ProductDTO productDTO) {
-            Product product = ProductMapper.INSTANCE.ProductDTOtoProduct(productDTO);
-            log.info("Saving product for user with id: {}", product.getUserId());
-            return productRepository.save(product);
+        Product product = ProductMapper.INSTANCE.ProductDTOtoProduct(productDTO);
+        log.info("Saving product for user with id: {}", product.getUserId());
+        return productRepository.save(product);
     }
 
     public Product getProductById(Long productId) {
-        return productRepository.findById(productId).orElseThrow(() -> new ProductNotFoundException(productId));
+        return productRepository.findProductById(productId);
     }
 
 
@@ -40,29 +43,37 @@ public class ProductService {
         return productRepository.findAll();
     }
 
-    public String deleteProduct(String productId) {
-        String responseMessage = String.format("Deleting product with id %s ", productId);
-        productRepository.deleteById(Long.parseLong(productId));
-        log.info("Deleting product with id: {}", productId);
-        return responseMessage;
+    public boolean deleteProduct(String productId,HttpServletRequest requestCurrentUser) {
+        Product currentProduct = getProductById(Long.parseLong(productId));
+        if (isEquals(currentProduct,requestCurrentUser)){
+            productRepository.deleteById(Long.parseLong(productId));
+            log.info("Deleting product with id: {}", productId);
+            return true;
+        }
+        log.error("wrong delete product: {}", productId);
+        return false;
     }
 
-    public String updateProduct(ProductDTO productDTO,String productId) {
+    public Product updateProduct(ProductDTO productDTO, String productId, HttpServletRequest requestCurrentUser) {
         Product currentProduct = getProductById(Long.parseLong(productId));
-        String responseMessage;
-        if (productDTO != null){
-
-            //ToDo сделать проверку на соответствие продукта и пользователя isEquals(){}
-
+        if (isEquals(currentProduct,requestCurrentUser)){
+            productDTO.setUserId(jwtFilter.getUserIdFromRequest(requestCurrentUser));
             currentProduct = ProductMapper.INSTANCE.ProductDTOtoProduct(productDTO);
-            responseMessage = String.format("Update product with title {}", productDTO.getTitle());
             log.info("Update product with title {}", productDTO.getTitle());
-            return responseMessage;
+            return productRepository.save(currentProduct);
         }
         return null;
     }
 
-//    Todo добавить метод "boolean isEquals(){}"
+    public boolean isEquals(Product currentProduct, HttpServletRequest request) {
+        long currentUserId = Long.parseLong(jwtFilter.getUserIdFromRequest(request));
+        long productOwnerId = Long.parseLong(currentProduct.getUserId());
+        return currentUserId == productOwnerId;
+    }
 
+    public List<Product> getAllProductFromUser(HttpServletRequest request) {
+        long userId = Long.parseLong(jwtFilter.getUserIdFromRequest(request));
+        return productRepository.findAllByUserIdOrderByCreateDateDesc(userId);
+    }
 
 }
